Create base classes for LDA/pLSA
